requirements.txt
peft==0.18.0
torch==2.9.1
transformers==4.57.3
pydantic>=2.0.0
pyyaml>=6.0
scikit-learn>=1.0.0
